🚚 Rename operations directory to operation
