+++ packages/core/src/index.ts
@@ -4,6 +4,6 @@ export * from './effect'
 export * from './executable-promise'
 export * from './generator'
 export * from './plugins'
-export * from './operations'
+export * from './operation'
 export { Task } from './stack'
 export * from './time'
