+++ packages/core/src/operation/terminal.ts
